Refactor into types file
