+++ src/Event/Event.js
@@ -12,6 +12,7 @@ import Animated, {
   useDerivedValue,
 } from 'react-native-reanimated';
 import styles, { circleStyles } from './Event.styles';
+import { EventPropType, EditEventConfigPropType } from '../utils/types';
 
 const DEFAULT_COLOR = 'red';
 const UPDATE_EVENT_ANIMATION_DURATION = 150;
@@ -303,23 +304,8 @@ const Event = ({
   );
 };
 
-export const EditEventConfigPropType = PropTypes.shape({
-  left: PropTypes.bool,
-  top: PropTypes.bool,
-  right: PropTypes.bool,
-  bottom: PropTypes.bool,
-});
-
-export const eventPropType = PropTypes.shape({
-  color: PropTypes.string,
-  id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
-  description: PropTypes.string,
-  startDate: PropTypes.instanceOf(Date).isRequired,
-  endDate: PropTypes.instanceOf(Date).isRequired,
-});
-
 Event.propTypes = {
-  event: eventPropType.isRequired,
+  event: EventPropType.isRequired,
   top: PropTypes.number.isRequired,
   left: PropTypes.number.isRequired,
   height: PropTypes.number.isRequired,
@@ -331,6 +317,7 @@ Event.propTypes = {
   onDrag: PropTypes.func,
   onEdit: PropTypes.func,
   editingEventId: PropTypes.number,
+  editEventConfig: EditEventConfigPropType,
 };
 
 export default React.memo(Event);
