+++ src/utils/types.js
@@ -1,0 +1,34 @@
+import PropTypes from 'prop-types';
+
+export const EditEventConfigPropType = PropTypes.shape({
+  left: PropTypes.bool,
+  top: PropTypes.bool,
+  right: PropTypes.bool,
+  bottom: PropTypes.bool,
+});
+
+export const EventPropType = PropTypes.shape({
+  color: PropTypes.string,
+  id: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
+  description: PropTypes.string,
+  startDate: PropTypes.instanceOf(Date).isRequired,
+  endDate: PropTypes.instanceOf(Date).isRequired,
+});
+
+export const EventWithMetaPropType = PropTypes.shape({
+  ref: EventPropType.isRequired,
+  box: PropTypes.shape({
+    startDate: PropTypes.instanceOf(Date).isRequired,
+    endDate: PropTypes.instanceOf(Date).isRequired,
+  }).isRequired,
+});
+
+export const GridRowPropType = PropTypes.shape({
+  borderColor: PropTypes.string,
+  borderTopWidth: PropTypes.number,
+});
+
+export const GridColumnPropType = PropTypes.shape({
+  borderColor: PropTypes.string,
+  borderLeftWidth: PropTypes.number,
+});
